Move message iterators to `stateful`
