+++ packages/engine/lib/stateful/Cargo.toml
@@ -9,6 +9,7 @@ memory = { path = "../memory" }
 arrow = { version = "10.0.0", default-features = false }
 lazy_static = "1.4.0"
 parking_lot = "0.11.1"
+rayon = "1.4.1"
 serde = { version = "1.0.111", features = ["derive"] }
 serde_json = "1.0.59"
 serde-aux = "0.6.1"
