+++ packages/engine/lib/stateful/src/message/arrow/mod.rs
@@ -1,2 +1,3 @@
 pub mod array;
 pub mod column;
+pub mod record_batch;
