+++ packages/engine/src/simulation/package/context/packages/api_requests/handlers.rs
@@ -1,11 +1,9 @@
 use serde_json::Value;
+use stateful::agent;
 use thiserror::Error as ThisError;
 
 use crate::{
-    datastore::{
-        table::{pool::message::MessageReader, references::AgentMessageReference},
-        UUID_V4_LEN,
-    },
+    datastore::{table::pool::message::MessageReader, UUID_V4_LEN},
     simulation::package::context::packages::api_requests::{
         handlers, ApiResponseMap, Error, Result,
     },
@@ -21,7 +19,7 @@ pub struct Requests {
 
 pub fn gather_requests(
     reader: &MessageReader<'_>,
-    messages: &[AgentMessageReference],
+    messages: &[agent::MessageReference],
 ) -> Result<Requests> {
     let inner = (0..messages.len())
         .into_iter()
